refactor(discord): nest slash commands under chat-input-command
Group slash-command UI, shared types/utils, and tests under chat-input-command,
update auto-import names, and fix knip/composer SSR assertions that were failing CI.
