+++ app/components/discord/chat-input-command/app-icon.vue
