+++ app/components/discord/chat-input-command/suggestions.vue
@@ -86,7 +86,7 @@
 								:aria-label="`${SlashCommandApps[app].label} commands`"
 								@click="selectedApp = app"
 							>
-								<DiscordSlashCommandAppIcon :app size="rail" />
+								<DiscordChatInputCommandAppIcon :app size="rail" />
 							</button>
 						</div>
 					</nav>
@@ -281,7 +281,7 @@ const ariaLabel = computed(() => `Slash command suggestions for ${prefix}`);
 }
 
 .discord-slash-command-suggestions-sidebar-recent {
-	/* Same 32px footprint as DiscordSlashCommandAppIcon rail so the active
+	/* Same 32px footprint as DiscordChatInputCommandAppIcon rail so the active
 	   squircle (40px / radius 12px) frames the clock like the app icons. */
 	@apply flex size-8 items-center justify-center max-md:size-7;
 }
