+++ shared/types/chat-input-command.ts
@@ -1,3 +1,33 @@
+type SlashCommandNameOnly = {
+	commandName: string;
+	subcommand?: undefined;
+	subcommandGroup?: undefined;
+};
+
+type SlashCommandDirectSubcommand = {
+	commandName: string;
+	subcommand: string;
+	subcommandGroup?: undefined;
+};
+
+export type SlashCommandGroupedSubcommand = {
+	commandName: string;
+	subcommandGroup: string;
+	subcommand: string;
+};
+
+export type SlashCommandDisplayParts =
+	| SlashCommandNameOnly
+	| SlashCommandDirectSubcommand
+	| SlashCommandGroupedSubcommand;
+
+export type SlashCommandDisplayInput = {
+	commandName: string;
+	subcommand?: string;
+	subcommandGroup?: string;
+	subcommandSubgroup?: string;
+};
+
 export interface SlashCommandOption {
 	name: string;
 	value?: string;
