+++ shared/types/index.ts
@@ -4,4 +4,4 @@ export * from "./discord";
 export * from "./env";
 export * from "./prisma";
 export * from "./utils";
-export * from "./slash-command";
+export * from "./chat-input-command";
